plugins: example of using the tee-supplicant plugin framework
TEE clients can add to the tee-supplicant own plugins.
This feature makes the supplicant a bit more flexible
in terms of providing services. Any external TEE services
can be designed as a tee-supplicant plugin.

User TAs can interact with the plugins using libutee
and the OP-TEE kernel code can use a special plugin PRC for it.

This patch adds example of using the plugin framework.
'syslog' plugin is used as an example.
It can write log messages from OP-TEE (TAs or kernel)
to system log.

Also patch adds a new TA, which interacts with the
'syslog' plugin with the help of 'tee_invoke_supp_plugin()'.
This TA increments a value and prints some strings
to the syslog. Also patch adds a host CA, which
calls the TA 'TA_PING_CNT' times.

If the example works successfully, we can find the following
strings in the log file (for qemu it's '/var/log/messages' file):
```
21:18:05 buildroot daemon.info tee-supplicant[92]: Hello, plugin! value = 0x0
21:18:07 buildroot daemon.info tee-supplicant[92]: Hello, plugin! value = 0x1
21:18:09 buildroot daemon.info tee-supplicant[92]: Hello, plugin! value = 0x2
21:18:11 buildroot daemon.info tee-supplicant[92]: Hello, plugin! value = 0x3
21:18:13 buildroot daemon.info tee-supplicant[92]: Hello, plugin! value = 0x4
```

Signed-off-by: Aleksandr Anisimov <a.anisimov@omprussia.ru>
Reviewed-by: Jens Wiklander <jens.wiklander@linaro.org>